+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Version 0.3.1 released 10/28/2023
-Added `freeAJSON` method that lets you recursively deallocate an aJSON struct.
+- Added `freeAJSON` method that lets you recursively deallocate an aJSON struct.
+- Added boolean property (char) `has_key` to the `aJSON` structure.
 
 # Version 0.3.0 released 10/21/2023
 Major release. Added new functions including optional C++ methods. See `README.md` for more information.